ディレクトリ分割、名前空間の変更
