+++ tests/Controller/Auth/LoginControllerTest.php
@@ -1,5 +1,5 @@
 <?php
-namespace Controller;
+namespace Controller\Auth;
 
 use Http\Http;
 use Model\User\Auth;
