+++ tests/Controller/Post/InsertControllerTest.php
@@ -1,44 +1,44 @@
 <?php
-namespace Controller;
+namespace Controller\Post;
 
 use Http\Http;
 use Model\Post\PostWriter;
 use PHPUnit\Framework\TestCase;
 
-class DeleteControllerTest extends TestCase
+class InsertControllerTest extends TestCase
 {
-	public function testDeleteAction()
+	public function testInsertAction()
 	{
 		$http = $this->getMockBuilder(Http::class)->getMock();
 		$http->expects($this->once())
 			->method('redirect')
-			->with('/user_page?user_id=1')
+			->with('/')
 		;
 
 		$post_writer = $this->getMockBuilder(PostWriter::class)
 			->disableOriginalConstructor()
 			->getMock();
 		$post_writer->expects($this->once())
-			->method('delete')
-			->with(1, 1)
+			->method('insert')
+			->with('test', 1)
 		;
 
-		$_SERVER['REQUEST_METHOD'] = 'POST';
+		$_POST['text'] = 'test';
 		$_SESSION['user_id'] = 1;
-		$_POST['id'] = 1;
-		$delete_controller = new DeleteController(
+		$_SERVER['REQUEST_METHOD'] = 'POST';
+		$insert_controller = new InsertController(
 			$http,
 			$post_writer
 		);
-		$delete_controller->deleteAction();
+		$insert_controller->insertAction();
 	}
 
-	public function testDeleteAction_NotPostRequest()
+	public function testInsertAction_NotPostRequest()
 	{
 		$http = $this->getMockBuilder(Http::class)->getMock();
 		$http->expects($this->once())
 			->method('redirect')
-			->with('/user_page?user_id=1')
+			->with('/')
 			->willReturnCallback(function () {
 				throw new \Exception('exit with redirect');
 			})
@@ -48,17 +48,16 @@ class DeleteControllerTest extends TestCase
 			->disableOriginalConstructor()
 			->getMock();
 		$post_writer->expects($this->never())
-			->method('delete')
+			->method('insert')
 		;
 
 		$_SERVER['REQUEST_METHOD'] = 'GET';
-		$_SESSION['user_id'] = 1;
-		$delete_controller = new DeleteController(
+		$insert_controller = new InsertController(
 			$http,
 			$post_writer
 		);
 		$this->expectException(\Exception::class);
 		$this->expectErrorMessage('exit with redirect');
-		$delete_controller->deleteAction();
+		$insert_controller->insertAction();
 	}
 }
